Move GROOT relative stats out of train script
